Better overflow handling. Test for overflow conditions.  Limit the size of grids, maps, tiles boards, etc. These limits help prevent overflows on large inputs.

+++ plat2d/draw.cc
@@ -24,8 +24,8 @@ int main(int argc, char *argv[]) {
 }
 
 static void parseargs(int argc, char *argv[]) {
-	for (int i = 1; i < argc; i++) {
-		if (strcmp(argv[i], "-o") == 0 && i < argc - 1) {
+	for (unsigned int i = 1; i < (unsigned int) argc; i++) {
+		if (strcmp(argv[i], "-o") == 0 && i < (unsigned int) argc - 1) {
 			outfile = argv[++i];
 		} else if (strcmp(argv[i], "-h") == 0) {
 			helpmsg(0);
